+++ puffin-imgui/src/ui.rs
@@ -1245,7 +1245,7 @@ fn max_memory_controls(ui: &Ui<'_>, frames: &Frames, frame_view: &GlobalFrameVie
     ));
 
     let frames_per_second = if let (Some(first), Some(last)) = (uniq.first(), uniq.last()) {
-        let nanos = last.meta.range_ns.1 - first.meta.range_ns.0;
+        let nanos = last.range_ns().1 - first.range_ns().0;
         let seconds = nanos as f64 * 1e-9;
         let frames = last.frame_index() - first.frame_index() + 1;
         frames as f64 / seconds